+++ src/multiple_stonks_v1.py
@@ -2,7 +2,7 @@ import json
 
 import openpyxl as pyxl
 
-import utils
+import file_handler
 
 from excel import openpyxl_helper
 from stonks import base_api, financial_data
@@ -127,7 +127,7 @@ def run_local():
 
     filename = '_multiple_stonks.xlsx'
     workbook.save(filename)
-    utils.open_file(filename)
+    file_handler.open_file(filename)
 
 if __name__ == '__main__':
     run_local()
